feat: subscribeInfo로 이름 변경 + K 포맷팅 유틸 함수 추가
